+++ libraries/BTstackLib/examples/iBeacon/iBeacon.ino
@@ -1,4 +1,4 @@
-#include <BTstack.h>
+#include <BTstackLib.h>
 #include <stdio.h>
 #include <SPI.h>
