+++ libraries/BTstackLib/library.properties
